Move checkBackupOperationBlocksAction into common_backup.go
It depends on GetOpenStackBackupOperationInProgress.
