+++ api/v1beta1/common_webhook.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/openstack-k8s-operators/osp-director-operator/api/shared"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	goClient "sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -66,22 +65,4 @@ func getRoleNames(namespace string) (map[string]string, error) {
 	}
 
 	return found, nil
-}
-
-func checkBackupOperationBlocksAction(namespace string, action shared.APIAction) error {
-	op, err := GetOpenStackBackupOperationInProgress(webhookClient, namespace)
-
-	if err != nil {
-		return err
-	}
-
-	if action == shared.APIActionCreate && (op == shared.BackupCleaning || op == shared.BackupSaving || op == shared.BackupReconciling) {
-		// Don't allow creation of certain OSP-D-operator-specific CRDs during backup save, (restore) clean or (restore) reconcile
-		err = fmt.Errorf("OSP-D operator API is disabled for creating resources while certain backup operations are in progress")
-	} else if action == shared.APIActionDelete && (op == shared.BackupLoading || op == shared.BackupSaving || op == shared.BackupReconciling) {
-		// Don't allow deletion of certain OSP-D-operator-specific CRDs during backup save, (restore) load or (restore) reconcile
-		err = fmt.Errorf("OSP-D operator API is disabled for deleting resources while certain backup operations are in progress")
-	}
-
-	return err
 }
